+++ integreat_cms/cms/urls/protected.py
@@ -767,7 +767,6 @@ urlpatterns = [
                                             "bulk-archive/",
                                             bulk_action_views.BulkArchiveView.as_view(
                                                 model=Page,
-                                                field_name="explicitly_archived",
                                             ),
                                             name="bulk_archive_pages",
                                         ),
@@ -775,7 +774,6 @@ urlpatterns = [
                                             "bulk-restore/",
                                             bulk_action_views.BulkRestoreView.as_view(
                                                 model=Page,
-                                                field_name="explicitly_archived",
                                             ),
                                             name="bulk_restore_pages",
                                         ),